Fix #2 - Rename sensor_samples to api_samples
